Ajeitando CSS, quantos pokemon, adicionando nome
Ajeitando CSS, agora mostra os pokemon de acordo com a quantidade que o usuário passa, adicionando o nome dos pokemon nas divs de cada um, botão de limpar também adicionando.
