+++ index.html
@@ -20,9 +20,10 @@
 
         <div id="divinputsEscolherPokemon">
 
-            <label for="escolherPokemon">Escolha o pokemon que deseja ver: </label>
-            <input type="text" id="escolherPokemon" name="escolherPokemon">
-            <input type="button" value="Aparecer Pokemon" id="botaoEscolherPokemon">
+            <label class="itemsHeader" for="escolherPokemon">Quantos Pokemon deseja ver?</label>
+            <input class="itemsHeader" type="text" id="escolherPokemon" name="escolherPokemon">
+            <input class="itemsHeader" type="button" value="Aparecer Pokemon" id="botaoEscolherPokemon">
+            <input class="itemsHeader" type="button" value="Limpar" id="botaoLimparPokemon">
 
         </div>
