+++ script.js
@@ -1,4 +1,5 @@
-import * as module from "./modules.js" 
+import * as module from "./modules.js"
+export {divAparecerPokemon, escolherPokemon}
 
 let escolherPokemon = document.querySelector("#escolherPokemon")
 
@@ -6,29 +7,8 @@ let botaoEscolherPokemon = document.querySelector("#botaoEscolherPokemon")
 
 let divAparecerPokemon = document.querySelector("#divAparecerPokemon")
 
-botaoEscolherPokemon.addEventListener('click', aparecerPokemon)
+let botaoLimparPokemon = document.querySelector("#botaoLimparPokemon")
 
-async function aparecerPokemon() {
+botaoEscolherPokemon.addEventListener('click', module.aparecerPokemon)
 
-    const resposta = await fetch(`https://pokeapi.co/api/v2/pokemon/${escolherPokemon.value}/`)
-
-    if(resposta.status == 404){
-        console.log("error")
-        return 0
-    }
-
-    const dados = await resposta.json()
-
-    console.log(dados)
-
-    let divPokemonEspecifico = document.createElement('div')
-
-    divPokemonEspecifico.classList.add("divPokemonEspecifico")
-
-    let imagemPokemon = module.adicionarImagem(dados.sprites.front_default)
-
-    divPokemonEspecifico.append(imagemPokemon)
-
-    divAparecerPokemon.appendChild(divPokemonEspecifico)
-
-}
+botaoLimparPokemon.addEventListener('click', module.limparPokemon)
