+++ style.css
@@ -23,24 +23,43 @@ h1{
     display: flex;
     justify-content: center;
     align-items: center;
+    padding: 0 1em;
     background-color: black;
     color: white;
     height: 5em;
 
 }
 
-#escolherPokemon{
-
-    margin: 0 3em;
-
+.itemsHeader{
+    margin: 0 1em;
 }
 
 #divAparecerPokemon{
-    display: flex;
-    justify-content: space-around;
+   display: flex;
+   flex-wrap: wrap;
+   justify-content: center;
 }
 
 .divPokemonEspecifico{
+    flex: 0 1 30%;
+    box-sizing: border-box;
+    margin: 1em;
     border: 0.2em solid;
     border-radius: 1em;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+}
+
+.h1NomePokemon{
+
+    font-size: 1.5em;
+
+}
+
+.h1NomePokemon::first-letter{
+
+    text-transform: uppercase;
+
 }
